+++ wallet-service/src/main/java/com/programming/techie/repository/LedgerRepository.java
@@ -4,8 +4,10 @@ import com.programming.techie.entity.LedgerEntry;
 import org.springframework.data.jpa.repository.*;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
+
 public interface LedgerRepository extends JpaRepository<LedgerEntry, Long> {
 
     @Query("SELECT COALESCE(SUM(l.amount),0) FROM LedgerEntry l WHERE l.walletId = :walletId")
-    Double calculateBalance(@Param("walletId") Long walletId);
+    BigDecimal calculateBalance(@Param("walletId") Long walletId);
 }
